[Refactor]GameState to transient and persistent storage
this way I can manage the lifetimes of memory I expect to live for the
duration of the game (persistent) and transient storage. I think this
also sets me up for an easier time to setup hot reloading.

Also changed the if statements to a switch statement in render code